[sycl] CLUE 3D SYCL implementation (#17)
* Ignored dev and data

* [sycl] updated sycl path and eigen dependencies

* Fixed md5

* [sycl] Fixed data import

Test compiles and runs

* [sycl] started working on CLUE implementation

* [sycl] actually fixed data import

* [sycl] removed test plugins

* [sycl] Implemented PointsLoader plugin

This plugin allocates memory on the device and fills it with data from the specified input file

* Moved location of input data for CLUEAlgo

* Added validator plugin to validate output of LoaderPlugin

* Updated for data

* Ignored sycl executable

* Added Support for nvidia gpus in patatrack02

* [sycl] Fixed validator for data on device memory

* [sycl] added ClueAlgoSYCL and dependencies

This plugin doesn't yet consume nor produce anything

* [sycl] added first version of CLUE clusterizer

The plugin will not compile with this producer at this point in time

* Revert "[sycl] added first version of CLUE clusterizer"

This reverts commit 7b1e4c4e54c50db8ecc3c11e620beeae882e9464.

* [sycl] added first version of CLUE clusterizer

The plugin will not compile with the clusterizer at this point in time

* [sycl] updated dependencies for CLUE clusterizer

* Fix CLUEClusterizer compilation due to datatypes

* Fixed segfault on view, now clusterizer runs

* [sycl] removed unused method

* [sycl] moved LayerTilesSYCL in SYCLDataFormats

* [sycl] modified producer to only read from input

* [sycl] updated producer name

* [sycl] moved back to unique pointers

* [sycl] added namespace cms for consistency

* [sycl] implemented copying and clustering
both operation now happen a single plugin

* [sycl] modified validator to only check input data

* [sycl] removed old and unused data format

* [sycl] added logic to choose device at runtime

* [sycl] added flags for AOT compilation

* [sycl] modified kernels to use only one dimension

* [sycl] fixed environment for choosing devices

* Output to file from CLUE output

* Updated to output CLUE results from correct Plugin Product

* [sycl] slightly updated copy validator

* [sycl] updated output producer
added custom constructor to PointsCloud
removed unsued data formats

* [sycl] updated dependencies for output producer

* [sycl] added output producer modules
(they are now under transfer)

* ToyDetector automatic validation with 100k points file coded and passing

* Validate with 10k toy detector

* Now output data dir is created automatically

* [sycl] cleaned up CLUE validator
Transfer back to host and csv creation are now disentagled
(csv production requires explicit passing of an extra parameter)

* Added reference output files

* Fixed input file corruption

* Modified input data filenames to avoid md5 issues

* Fixed alpaka checkout commit

* Fixed dataclean

* Added target to clear output files

* [sycl] implemented runtime choice for input file

* [sycl] plugins read from specified input file

* [sycl] validators now check the correct file
Result check for toyDetectors happen automatically (if --validation)

* Ignores reference data output

* [sycl] moved to initializaition list

* [sycl] added options in the help message

* One makefile for building in patatrack or olice depending on variable 'USE_SYCL_PATATRACK'

* Make CLUEValidator ONLY validate the output of CUE

* Added test config files

* [sycl] added config file choice
Clusterizer now has an ESProducer
Modified output file name creation to include parameters

* [sycl] removed test output

* [sycl] moved PluginWrapper in cms::sycltools

* [sycl] fixed help text

* [sycl] more fixes for help text

* Updated with intructions and examples
Mainly specific to sycl implementation

* [sycl] implemented framework atomics

* [sycl] Reverted atomics
New form is currently incompatible with CUDA backend

* [sycl] improved transfer logic and code clean up

* [sycl] updated README

* [sycl] removed reduntant assignment

* Ignored performance plots

* Moved CLUE files in a single archive
Removed eigen library

* [sycl] removed reduntant include

* [sycl] test now ignores FPGAs
Can now be run without errors with AOT compilation

* CLUE data is now in a single archive
Removed Pixeltrack data

* [sycl] Removed dependency from Pixeltrack data
Still relies on plugin-Pointloader to get PointsCloud to cluster

* [sycl] Updated help message

* Ignored config and dev

* [sycl] Updated test to not rely on Pixeltrack data

* [sycl] removed need for PointsLoader
Points are now loaded into host memory by Source

* [sycl] Clusterizer doesn't need loader anymore

* [sycl] Simplified validation in a single plugin
Now automatically checks if parameters are the same as reference
Still checks input variables for all toyDetectors

* Added table of contents and SYCL device selection

* [sycl] Fixed input loading in Source

* [sycl] Streamlined output production
Less copies required as input data is kept the same

* [sycl] Fixed source to read all the layers
One complete input file is now one event

* Performance Measuring Script

* [sycl] moved data reading to binary

Added calculation of maximum throughput based on reading speed
Measures maxEvents reads of the same file to simulate  more events

* [sycl] Implemented tiles selection on Intel GPUs

* Updated config files

* [sycl] Prepared to store input data in one binary

* [sycltest] moved to binary data reading

* Updated url and md5 with new data

* Updated all developed backend to use new input data

* [sycl] improved device selection code

* [sycltest] improved device assignation logic

* [sycl] Added caching allocator

* [sycl] Implemented caching allocator in CLUEAlgo

* [sycl] improved caching allocator performance

* [sycl] Added debug flag for device selector

* [sycl] added documentation for caching allocator

* [sycl] Add caching allocator policy selection
It can now be disabled using SYCL_DISABLE_CACHING_ALLOCATOR

* [sycltest] added caching allocator in SYCLCore

* General fixes for all backends

* Fixed remote link for data

* [sycl] implemented host caching allocator

* [sycl] caching allocator now checks the context

* [sycl] one host caching allocator per platform

- Device caching allocators are virtually unchanged
- Solidified caching allocators' interface

* [sycl] Updated atomics

* Now all implementations use oneTBB from cvmfs

* Fixed TBB location in oneAPI base

* [sycl] updated caching allocator debug information

* [sycl] updated device selection code

* [sycl] added EventProcessor and Source for CLUE3D

* [sycl] added CLUE3D DataFormats

* [serial] [alpaka] fixed type of a parameter

* [sycl] updated VecArray and added printf

Moved VecArray to SYCLCore to conform with other repos

* [sycl] implemented CLUE3D

* [sycl] increase threads per block from 256 to 1024

Virtually no change on CPU performance, GPUs seem to run better

* [serial] # tracksters is not printed by default

* [sycl] Added llvm chain for HIP and CUDA

* [sycl] results are now printed using a sycl stream

printf was not compatible with hip architecture

* [sycl] added CPU target in LLVM toolchain

* Optimized CLUE2D following Alpaka implementation

+++ src/alpaka/plugin-CLUETracksterizer/alpaka/CLUE3DAlgoKernels.h
@@ -54,7 +54,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
                                   pointsView *d_points,
                                   uint32_t const &numberOfPoints,
                                   int densitySiblingLayers = 3,
-                                  int densityXYDistanceSqr = 3.24,
+                                  float densityXYDistanceSqr = 3.24,
                                   float kernelDensityFactor = 0.2,
                                   bool densityOnSameLayer = false) const {
       int nEtaBin = ticl::TileConstants::nEtaBins;
